+++ src/Draft/Convert-HashTableToVariable.ps1
@@ -1,0 +1,5 @@
+function Convert-HashtableToVariable {
+   [CmdletBinding(HelpUri = 'https://day3bits.com/PSPreworkout')]
+   param ()
+   foreach ($key in $_.Keys) { Set-Variable -Name $key -Value $_[$key] }
+}