+++ src/PSPreworkout/Public/Get-TypeAccelerator.ps1
@@ -9,31 +9,35 @@ function Get-TypeAccelerator {
     .PARAMETER Name
     The name of a specific type accelerator to get.
 
-    .PARAMETER GridView
-    Show the output in a grid view.
-
     .EXAMPLE
     Get-TypeAccelerator -Name ADSI
 
+    .EXAMPLE
+    Get-TypeAccelerator -Name ps*
+
+    Get all type accelerators that begin with the string "ps".
+
     .NOTES
+    Author: Sam Erde
+    Version: 0.0.3
+    Modified: 2024/10/29
+
     Thanks to Jeff Hicks (@JDHITSolutions) for helpful suggestions and improvements on this output!
 
+    Change Log: Removed the grid view option to allow user flexibility in how they want to output the results.
+    To Do: Add a way to filter/search the 'type' property.
+
     #>
-    [CmdletBinding()]
+    [CmdletBinding(HelpUri = 'https://day3bits.com/PSPreworkout/Get-TypeAccelerator')]
+    [OutputType('System.Array')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseConsistentIndentation', '', Justification = 'Fighting with VS Code autoformatting.')]
-    #[System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', 'Get-TypeAccelerators', Justification = 'The type accelerators are plural.')]
     param (
 
         # Parameter help description
         [Parameter(Position = 0, HelpMessage = 'The name of the type accelerator, such as "ADSI."')]
         [SupportsWildcards()]
         [string]
-        $Name = '*',
-
-        # Show a grid view of the loaded assemblies
-        [Parameter()]
-        [switch]
-        $GridView
+        $Name = '*'
     )
 
     $TypeAccelerators = ([PSObject].Assembly.GetType('System.Management.Automation.TypeAccelerators')::Get).GetEnumerator() |
@@ -47,22 +51,5 @@ function Get-TypeAccelerator {
                     Type       = $_.Value.FullName
                 }
             }
-
-    if ($PSBoundParameters.ContainsKey('GridView')) {
-
-        if ((Get-Command -Name Out-ConsoleGridView -ErrorAction SilentlyContinue) -and ($PSVersionTable.PSEdition -eq 'Core')) {
-
-            $TypeAccelerators | Out-ConsoleGridView -OutputMode Multiple
-
-        } elseif ((Get-Command -Name Out-GridView -ErrorAction SilentlyContinue) -and ($PSVersionTable.PSEdition -eq 'Desktop')) {
-
-            $TypeAccelerators | Out-GridView -OutputMode Multiple
-
-        } else {
-            Write-Output 'The Out-GridView and Out-ConsoleGridView cmdlets were not found. Please install the Microsoft.PowerShell.ConsoleGuiTools module or re-install the PowerShell ISE if using Windows PowerShell 5.1.'
-            $TypeAccelerators | Format-Table -AutoSize
-        }
-    }
-
     $TypeAccelerators
 }