+++ src/PSPreworkout/Public/Install-OhMyPosh.ps1
@@ -32,7 +32,7 @@ function Install-OhMyPosh {
     .EXAMPLE
         Install-OhMyPosh
 #>
-    [CmdletBinding()]
+    [CmdletBinding(HelpUri = 'https://day3bits.com/PSPreworkout/Install-OhMyPosh')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingInvokeExpression', '', Justification = 'Invoke-Expression is used for online OMP installer.')]
     param (
         [Parameter()]