+++ src/PSPreworkout/Public/Set-EnvironmentVariable.ps1
@@ -20,7 +20,7 @@ function Set-EnvironmentVariable {
 
     #>
     [Alias('sev')]
-    [CmdletBinding()]
+    [CmdletBinding(HelpUri = 'https://day3bits.com/PSPreworkout/Set-EnvironmentVariable')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '', Justification = 'OK')]
     param (
         # The name of the environment variable to set.