+++ src/PSPreworkout/Public/Update-AllTheThings.ps1
@@ -27,7 +27,10 @@ function Update-AllTheThings {
     Updates all of the things it can!
     #>
 
-    [CmdletBinding(SupportsShouldProcess)]
+    [CmdletBinding(
+        SupportsShouldProcess,
+        HelpUri = 'https://day3bits.com/PSPreworkout/Update-AllTheThings'
+    )]
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', 'Update-AllTheThings', Justification = 'This is what we do.')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingWriteHost', '', Justification = 'Interactive Use')]
     [Alias('uatt')]
@@ -70,7 +73,7 @@ function Update-AllTheThings {
 /_  __/ /  ___   /_  __/ /  (_)__  ___ ____
  / / / _ \/ -_)   / / / _ \/ / _ \/ _ `(_-<
 /_/ /_//_/\__/   /_/ /_//_/_/_//_/\_, /___/
-                                 /___/ v0.5.7
+                                 /___/ v0.5.9
 
 "@
         Write-Host $Banner
@@ -195,7 +198,7 @@ function Update-AllTheThings {
 
         #region UpdateWinget
         # >>> Create a section to check OS and client/server OS at the top of the script <<< #
-        if ($IsWindows -or ($PSVersionTable.PSVersion -ge [version]'5.1')) {
+        if ($IsWindows -or ($PSVersionTable.PSVersion -le [version]'5.1')) {
 
             if ((Get-CimInstance -ClassName CIM_OperatingSystem).Caption -match 'Server') {
                 # If on Windows Server, prompt to continue before automatically updating packages.
@@ -252,6 +255,10 @@ function Update-AllTheThings {
                 Write-Host '[5] Updating apt packages.'
                 sudo apt update
                 sudo apt upgrade
+            }
+            if (Get-Command dnf -ErrorAction SilentlyContinue) {
+                Write-Host '[5] Updating dnf packages.'
+                sudo update
             }
         } else {
             Write-Verbose '[5] Not Linux. Skipping section.'