+++ CHANGELOG.md
@@ -1,18 +1,134 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 &nbsp;
+
+## [1.5.0] - 2024-11-04
+
+This release is focused on enhancements that make the existing functions more usable and flexible. The main changes are refactoring Get-EnvironmentVariable and Out-JsonFile.
+
+### Changed
+
+- Improve `Get-EnvironmentVariable`. Return rich objects to pipeline and test more usage scenarios.
+- Remove requirement for ".json" extension in `Out-JsonFile`.
+- Accept input from pipeline in `Out-JsonFile`.
+- Ensure file is created in file system provider with `Out-JsonFile`.
+- Remove **GridView** parameters from `Get-TypeAccelerator` to give user control of output and display.
+
+&nbsp;
 &nbsp;
 
-## [Unreleased]
+## [1.4.9] - 2024-10-25
+
+### Fixed
+
+- Fix [#22](https://github.com/SamErde/PSPreworkout/issues/22) in `Edit-WinGetSettingsFile` by temporarily removing the **EditorPath** parameter. It will be back! (@SamErde, @JDHITSolutions)
 
 ### Added
 
-- Add `Get-VSCodePortable`
-- Add `Remove-OldModule`
+- Add basic support for DNF on Linux to `Update-AllTheThings`. (@x1101)
+
+&nbsp;
+&nbsp;
+
+## [1.4.8] - 2024-10-24
+
+### Fixed
+
+- Fix [#26](https://github.com/SamErde/PSPreworkout/issues/26) Windows/Linux detection in `Update-AllTheThings`. (@x1101)
+
+&nbsp;
+&nbsp;
+
+## [1.4.7] - 2024-10-23
+
+### Changed
+
+- Fix HelpUri to open help pages to each specific function when `-Online` help is requested.
+
+&nbsp;
+&nbsp;
+
+## [1.4.6] - 2024-10-22
+
+### Added
+
+- Add updatable and online [external] help.
+
+&nbsp;
+&nbsp;
+
+## [1.4.3] - 2024-10-16
+
+### Fixed
+
+- Fix object creation example for comment-based help in `Show-WithoutEmptyProperty`
+
+&nbsp;
+Full changelog: [v1.4.2...v1.4.3](https://github.com/SamErde/PSPreworkout/compare/v1.4.2...v1.4.3)
+&nbsp;
+&nbsp;
+
+## [1.4.2] - 2024-10-15
+
+### Changed
+
+- Refactor `Show-WithoutEmptyProperty` and test with more examples
+- Enhance `Out-JsonFile` with better validation and provider handling
+
+&nbsp;
+Full changelog: [v1.4.1...v1.4.2](https://github.com/SamErde/PSPreworkout/compare/v1.4.1...v1.4.2)
+&nbsp;
+&nbsp;
+
+## [1.4.1] - 2024-10-14
+
+### Fixed
+
+- Fix [#16](https://github.com/SamErde/PSPreworkout/issues/16)
+- Fix [#15](https://github.com/SamErde/PSPreworkout/issues/15)
+
+&nbsp;
+Full changelog: [v1.4.0...v1.4.1](https://github.com/SamErde/PSPreworkout/compare/v1.4.0...v1.4.1)
+&nbsp;
+&nbsp;
+
+## [1.4.0] - 2024-10-14
+
+### Added
+
+- Add `Install-CommandNotFoundUtility`
+- Add preferred `EditorPath` parameter to `Edit-WinGetSettingsFile`
+
+### Fixed
+
+- Fix download links for `Get-PowerShellPortable`
+- Optimize `Get-PowerShellPortable` version check and download
+
+&nbsp;
+Full changelog: [v1.3.1...v1.4.0](https://github.com/SamErde/PSPreworkout/compare/v1.3.1...v1.4.0)
+&nbsp;
+&nbsp;
+
+## [1.3.1] - 2024-10-10
+
+### Added
+
+- Add `Edit-PSReadLineHistoryFile`
+- Add `Out-JsonFile`
+- Add `Show-WithoutEmptyProperty`
+
+### Fixed
+
+- Fix broken build output due to corrupt formatter 🤷‍♂️
+- Comment out incomplete parameters in `Initialize-PSEnvironmentConfiguration`
+- Remove `-Force` on module install scriptblock for `Initialize-PSEnvironmentConfiguration`
+
+### Removed
+
+- Remove user installation of PowerShell, which no longer provides user installation packages after 7.2.6
 
 &nbsp;
 &nbsp;
@@ -20,9 +136,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [1.2.5] - 2024-09-16
 
 ### Changed
+
 - Set ~/Downloads as the default download location for all OSes
 
 ### Fixed
+
 - Fix some minor errors in path processing
 
 &nbsp;
@@ -31,8 +149,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [1.2.4] - 2024-09-13
 
 ### Fixed
+
 - Fix extraction of PowerShell .tar.gz file on macOS
-- Fix/improve references to location of the extracted PowerShell files
+- Fix references to location of the extracted PowerShell files
 
 &nbsp;
 &nbsp;
@@ -48,7 +167,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 &nbsp;
 &nbsp;
 
-
 ## [1.1.2] - 2024-09-12
 
 ### Changed
@@ -58,7 +176,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 &nbsp;
 &nbsp;
-
 
 ## [1.1.0] - 2024-09-11
 
@@ -78,7 +195,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 &nbsp;
 &nbsp;
 
-
 ## [0.4.1] - 2024-09-09
 
 ### Changed
@@ -87,7 +203,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 &nbsp;
 &nbsp;
-
 
 ## [0.3.2] - 2024-09-06
 
@@ -107,10 +222,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 _Initial release._
 
-[Unreleased]: https://github.com/SamErde/PSPreworkout/compare/v1.2.2...HEAD
+[1.4.9]: https://github.com/SamErde/PSPreworkout/tag/v1.4.9
+[1.4.8]: https://github.com/SamErde/PSPreworkout/tag/v1.4.8
+[1.4.7]: https://github.com/SamErde/PSPreworkout/tag/v1.4.7
+[1.4.6]: https://github.com/SamErde/PSPreworkout/tag/v1.4.6
+[1.4.3]: https://github.com/SamErde/PSPreworkout/tag/v1.4.3
+[1.4.2]: https://github.com/SamErde/PSPreworkout/tag/v1.4.2
+[1.4.1]: https://github.com/SamErde/PSPreworkout/tag/v1.4.1
+[1.4.0]: https://github.com/SamErde/PSPreworkout/tag/v1.4.0
+[1.3.1]: https://github.com/SamErde/PSPreworkout/tag/v1.3.1
 [1.2.5]: https://github.com/SamErde/PSPreworkout/tag/v1.2.5
 [1.2.4]: https://github.com/SamErde/PSPreworkout/tag/v1.2.4
-[1.2.3]: https://github.com/SamErde/PSPreworkout/tag/v1.2.3
 [1.2.2]: https://github.com/SamErde/PSPreworkout/tag/v1.2.2
 [1.1.2]: https://github.com/SamErde/PSPreworkout/tag/v1.1.2
 [1.1.0]: https://github.com/SamErde/PSPreworkout/tag/v1.1.0